refactor: refactor patient service
