+++ breeze/services/patient_service/appointment.py
@@ -1,0 +1,2 @@
+def manage_appointment(user, auth_service):
+    pass
